feat(github): M17-3 PR authoring entry — shared PullRequestSheet + mailbox toolbar (#192)
Extract the M16-3 PullRequestSheet (private in RemoteMailboxView.swift)
into PullRequestSheet.swift, internal, byte-for-byte the same flow
(push-first when no upstream via the M16-2 one-shot → POST /pulls →
open in browser). The Remote mailbox presents it unchanged; the Pull
Requests mailbox toolbar gains New Pull Request… presenting the same
sheet, refreshing the list on success so the new PR appears. No new
git verbs, no new API calls, no new auth surface — pure relocation +
one toolbar entry.
